인터페이스 추가 및 에셋추가 기타 코드 수정 및 아이템 줍기 기능 추가중

+++ Source/TeamNYC/Interfaces/InteractionInterface.h
@@ -44,7 +44,7 @@ struct FInteractableData
 
 	// used only for pickup
 	UPROPERTY(EditInstanceOnly)
-	int8 Quantity;
+	int32 Quantity;
 
 
 	// used for things like valves, doors, etc. that require an interaction timer
@@ -58,14 +58,9 @@ class UInteractionInterface : public UInterface
 	GENERATED_BODY()
 };
 
-/**
- * 
- */
 class TEAMNYC_API IInteractionInterface
 {
 	GENERATED_BODY()
-
-	// Add interface functions to this class. This is the class that will be inherited to implement this interface.
 public:
 	virtual void BeginFocus();
 	virtual void EndFocus();
